PowerShellCmd: improve InterruptedException
We need to wait for the sub-process to exit, then wait for the two
stream reader threads to finish. Handle InterruptedException from
any of these in a single loop.

Also, deal with the gnarly inconsistency where if p.waitFor() does
get interrupted, then it throws InterruptedException without clearing
the interrupted flag. This means every successive call to waitFor()
would also throw the exception unless we explicitly clear the flag.