Basic YAML marshalling support
Looks like:

  --- !vm
  actions: !actions
    links:
      - !link
        href: "http://{server}/vms/3/shutdown"
        rel: shutdown
        type: application/xml
      ...
  id: "{id}"
  link: !link
    href: "http://{server}/vms/{id}"
    rel: self
    type: application/xml

Still need to e.g. construct <link>s with application/x-yaml when
appropriate.

Also need to think about how to expose the top-level links with a
yaml media type.

Finally, need to think about this:

  transfers:
    vm: com.redhat.rhevm.api.dummy.DummyVM

the jyaml config lives in the api package, which doesn't know anything
about DummyVM. Can we somehow get it to marshal the object as a VM
instead of a DummyVM?