Fix yaml subclass marshalling issue
Major suckage alert!

We want to marshal only the common part of our entities - e.g. if
we return a DummyVM, we actually only want to marshal it as a VM
object.

There doesn't seem to be a trivial way to fix this, so for now add lame
copy constructors to the base entity classes. Perhaps we can work around
this by implenting our own org.ho.yaml.wrapper.ObjectWrapper?