Make powershell addLinks() methods non-static
Using ActionsBuilder in addLinks() will mean we need an actual
instance of the resource with the current state of the object
in order to check which actions are valid.

The only tricky bit is the changes to the host and vm resource
'prototype' code - we've lost the synchronization around the
cache of resource state, but we need further thought here
because we do the likes of:

  state = runPowerShellCommand(cmd);

  addLinks();

  return state;

If we can have multiple requests at once, we can be returning
state with no links or with links which were built using the
context of another request.