Implement powershell storagedomain initialize and teardown
The tricky part here is that it's only after we add the storage
domain in RHEV-M during initialize do we know its real ID.

So, we allocate an ID when the storage domain is created by POST
and maintain a mapping of that ID to RHEV-M's ID.

Note, this mapping is currently only in memory, so the IDs will
change if the server restarts. To fix this, we'll need to add
these ID mappings to a permanent database.