Fix NullPointerException in PowerShellStorageDomainResource.validateAction()
If we do a GET on a storage domain and we don't have the resource
cached, we do the following:

  1) in createSubResource() create a model with only an ID

  2) in get() run 'get-storagedomain' and parse the output

  3) pass the result of this to addLinks()

  4) addLinks() use an action builder to add action links

  5) the action builder calls validateAction()

  6) we eventually do setModel() with the result of all this

Unfortunately, we get an NPE in (5) because it expects the
model to be fully up to date with the latest status of the
resource.

Fix by doing setModel() before using the action builder.