Use storage domain Status/SharedStatus properties as appropriate
Status is the correct one to use when querying a storage domain in the
context of a data center - i.e. the status on an attachment.

SharedStatus is appropriate when querying a storage domain outside of
the context of a data center.

Because of this we need a class to represent the result of a powershell
command, as distinct from the state of the resource exposed by the
API. We do this by adding a PowerShellStorageDomain class which extends
the base StorageDomain class.