Added support for HTTP Basic auth.
Any incoming requests without an Authorization header are now met with a
WWW-Authenticate challenge.

If the Authorization header is present, its decoded and the principal is
cached in a ThreadLocal-based current for access further on in the dispatch
path. This Current-based approach is a work-around for RESTEASY-420, and will
be removed once the next release of RESTEasy is available.