Create templates using add-template -async
Adding a new template is guaranteed to take a long time because disk
images are copied. Rather than blocking the client for a long period
and perhaps timing out the TCP connection, we return a template
representation with a "LOCKED" status.

Luckily, this pretty much describes what the "-async" flag does, so
use that. Clearly, there are other commands we should use -async for
similar reasons.

There is some debate as to whether we should be returning 202 Accepted
rather than 200 Created in this case, but we can resolve that later.