Remove the storage domain initialize action
We initially avoided having a POST to the storage domains collection
actually create a storage domain for 2 reasons:

  1) You need to supply a host on which the storage domain initialization
     will take place, but the host association then disappears after
     the domain is created - i.e. you supply a <host> reference in
     the initial <storage_domain> and then the <host> disappears

  2) We can't supply a host parameter to DELETE, so we were going to
     need a "teardown" action anyway

  3) Initializing a storage domain is a long running operation, and
     at the time we felt that actions were the most suitable way of
     handling such operations

We've since decided we need to handle long-running resource creation
and had another example of (1) where adding a template needs a vm
reference.

Conclusion is that we should remove the initialize action, which
greatly reduces the complexity we had around staged storage domains
and ID mappings. We still have the teardown action, though.