Added support for transport-level encryption over HTTPS.
By default, the mock and powershell webapps are now deployed with a
security-constraint requiring confidentiality in the transport. This
generally requires further configuration in the servlet container to
enable (for example in the Tomcat case, uncommenting the "SSL HTTP/1.1
Connector" element in $CATALINA_HOME/conf/server.xml, also adding
appropriate keystoreFile and keystorePass attributes).

In order to build an insecure version of the webapps use:

 $ mvn install -Pinsecure

In the case of the embedded jetty engine, support is also provided
for enabling SSL and automatically generating a keystore containing
a self-signed certificate.

To run either webapp securely in jetty use:

 $ mvn jetty:run -Psecure

The keystore is generated in ./target/security/keystore and the
jetty connector is configured to listen for HTTPS requests on 8943.

Whereas to turn off transport-level security, use instead:

 $ mvn jetty:run -Pinsecure