Allow webapp secure and insecure profiles run in any order.
Previously running:

  $ mvn jetty:run -Pinsecure

before having ever activated the secure profile resulted in a
failure to find the configured keystore.

Now the SSL connector is only launched when the secure profile is
activated and vice versa for the plain ol' HTTP connector in the
insecure case (whereas previously both connectors were used in
either case).