powershell: add ConvertTo-Xml support in PowerShellCmd
Add a parameter to PowerShellCmd.runCommand() that causes the output
to be formatted using ConvertTo-Xml.

Note, we need to escape all single-quotes within the command before
wrapping it in single-quotes. It looks like the way to escape
single-quotes within a single-quoted string in powershell is with
another single-quote.

i.e. this:

  :> write-host 'foo''bar'
  foo'bar

as opposed to a unix shell's behaviour:

  $> echo 'foo''bar'
  foobar