Workaround for wrong content-type on device sub-collections.
Device sub-collections of a VM resource were marshalled as YAML if the
Accept header wasn't explicit specified in the corresponding request.
The @Produces annotation on the ReadOnlyDevicesResource interface should
have been sufficient to ensure marshalling as application/xml.

A fix to RESTEasy is pending but this work-around reslved the problem
in the meantime.