Add VM action to create a display ticket.
The action is named "ticket" as opposed to "new_ticket" or "create_ticket",
after the American usage of noun as verb.

The ticket value and/or expiry may be set by the caller, or left unset,
in which these are set internally.

The VM id is not directly exposed in the Ticket representation, as this is
implicit in the "parent" link.

See https://fedorahosted.org/rhevm-api/ticket/24