Switching work-around for PermGen starvation in powershell units.
Initially worked around PermGen OutOfMemoryError in the powershell tests
by setting the surefire forkMode to 'never'.

However, this approach is incompatible with maven parallel builds, hence
we switch to an alternative workaround, settin the -XX:MaxPermSize option
explicitly in the surefire configuration.

The underlying problem is suspected to be a surefire bug preventing the
-XX:MaxPermSize set in the MAVEN_OPTS being passed on to the forked JVM
with the default 'once' forkMode.