common: add support to LinkHelper for multiple parent types
Tag collections can have multiple parents - they can be a sub-resource
of host, user or VM resources as well as being a top-level collection.

Add support for this by replacing the existing parentType with a
map of parentType to collection definition.

This is a fairly significant refactor, but hopefully it makes this
code somewhat less fugly.