Switching to absolute paths in link hrefs, field-level injection of UriInfo.
  https://fedorahosted.org/rhevm-api/ticket/83

The previous convention that links paths were relative to the entry point URI
was considered in contravention of the relavent RFC:

  https://fedorahosted.org/pipermail/rhevm-api/2010-September/thread.html#305

So a link previous of form:

  <link rel="..." href="resource/{id}"/>

will now look like:

  <link rel="..." href="/rhevm-api-powershell/resource/{id}"/>

Note that any scripts that depend on the previous form will have to be updated.

Also the widespread use of parameter-level injection of UriInfo in the JAX-RS
resource interfaces has been replaced with field-level injection, leaving
these interfaces significantly cleaner.