powershell: always run logout-user before login-user
https://fedorahosted.org/rhevm-api/ticket/95

Powershell is often setup with a profile.ps1 script which automatically
logs the user into RHEV-M. We need to run logout-user to ensure that
those credentials are always dropped to make it safe to run the API
service as such a user.

Also, removal the conditional on principal being non-null/NONE - this
should never be possible since users will always be challenged for
credentials.