api: added Permit schema types and resources
A Permit captures the state of being allowed to carry out a set
of related operations on a particular entity type.

A Role in turn aggregates a set of Permits to give a coherent group of
operations that the Role allows.

The current set of supported Permits is exposed via the Capabilities
resource. Individual Permits are identified via an intergal ID and
also carry a descriptive name.