api: added Permission schema types and resources
A Permission is a tuple relating a User, a Role and an entity,
effectively capturing the notion of a role assigned to a particular
user with respect to an individual entity instance.

The permissions sub-collection of a UserResource models the set of
role assignments for that user with respect to any entity.

The permissions sub-collection of an entity resource models the set of
role assignments for any user with respect to that entity.