api: re-name the "isos" collection to "files"
https://fedorahosted.org/rhevm-api/ticket/81

ISO storage domains can also contain floppy files (.vfd).

Make the "isos" collection a little more general by re-naming it to
"files" and adding a "type" element which can be ISO or VDFD.

i.e.

  GET /rhevm-api/datacenters/123456/files HTTP/1.1
  Accept: application/xml

  HTTP/1.1 200 OK
  Content-Type: application/xml

  <files>
    <file id="foo.iso" href="/rhevm-api/datacenters/123456/files/foo.iso">
      <name>foo.iso</name>
      <type>ISO</type>
      <data_center id="123456" href="/rhevm-api/datacenters/123456"/>
    </file>
    <file id="bar.vfd" href="/rhevm-api/datacenters/123456/files/bar.vfd">
      <name>bar.vfd</name>
      <type>VFD</type>
      <data_center id="123456" href="/rhevm-api/datacenters/123456"/>
    </file>
  </files>

Note, the powershell CLI does not have a way of listing VFD files
so they won't be listed in the powershell implementation of the
API.