api: re-enable YAML support with a custom bean wrapper
https://fedorahosted.org/rhevm-api/ticket/1

We disabled YAML serialization back in the day because the YAML encoder
could not handle xjc generated getList() methods because it needs both
a getter and setter for each bean property.

Work around that issue by adding a custom bean wrapper which exposes
such properties.

As an example, this XML snippet:

  <vm id="12345" href="/rhevm-api/vms/12345">
    <stateless>true</stateless>
    <status>UP</status>
    <memory>4294967296</memory>
    <os>
      <kernel>vmlinuz</kernel>
      <boot dev="cdrom"/>
    </os>
  </vm>

Looks like this in YAML:

  !vm
  id : "12345"
  href : /rhevm-api/vms/12345
  stateless : true
  status : UP
  memory : 4294967296
  os: !os
    kernel: vmlinuz
    boot:
      - !boot:
         dev: CDROM

Note, however that serialization of enums and booleans doesn't currently
work that way - it needs further work.