powershell: make cdrom update work with running VM
Update to cdrom will now try to change the running config first, if
that fails (vm is not running) it will try to update the default config.

Note, update-vm is not tried first since it succeeds and does nothing
if the path is the same as the one in the default config.