api: add parameter to storage domain delete, drop the remove action
We've long struggled with the question of how to most sanely model
storage domain add and remove operations, given that both require
a host to be specified which will be used to initialize and re-format
the underlying physical storage.

Initially, we hoped that we could have a host be automatically
selected, but this isn't easy since not all hosts have access to
all storage.

Next, we took the approach of having initialize and teardown
actions because:

  1) We didn't like supply a host reference when creating a storage
     domain and then that host reference disappears from the storage
     domain representation

  2) We didn't think it would be appropriate to have any parameters
     required for the DELETE operation

  3) Since both add and remove are potentially long-running operations
     we felt both were best modelled as actions

Well, we've since gotten over (1) - we feel it's something that can
be lived with. So, we removed the initalize action.

Also, we've since realized that we need to handle long-running
creation operations using 202 Accepted.

Finally, on (2) it's become apparent that, while it's quite unorthodox,
to require a message body to DELETE, it's not actually prohibited
by the HTTP spec. Also, it appears to be supported fully by RESTEasy
under at least jetty and JBoss.

So, the conclusion is to drop the remove action, drop the TORNDOWN
storage domain status and require a parameter to DELETE that will
specify the host e.g.

  DELETE /rhevm-api/storagedomains/123456 HTTP/1.1
  Content-type: application/xml

  <storage_domain>
    <host id="654321"/>
  </storage_domain>

  HTTP/1.1 204 No Content