api: support for asynchronous resource creation
Certain resources (e.g. Disks, Snapshots, VMs, Templates) may be
created asynchronously by the RHEVM backend.

The schema types now allow creation status to be exposed in the initial
resource returned from the POST operation on the containing collection
so that the client may discern whether the resource creation is pending,
in progress, or complete.

A link with rel="creation_status" will also be included in this initial
representation, so that the client can monitor the progress of the
asynchronous creation before using the resource.

A future commit will add support for allowing the client to force
blocking behaviour (overriding the natural asynchrony of the creation)
via the "Expect: 201-created" header.