powershell: add injection of HttpHeaders
The HttpHeaders are injected by the JAX-RS runtime in the case of
the top level Spring-loaded collections resources.

For sub-collections and sub-resources requiring access to the headers,
the corresponding top-level resource must inject on creation.

We also provide a mechanism for the Expect header to be used to
express a client requirement that a potentially asynchronous resource
creation be guaranteed blocking (by specifying "Expect: 201-creation"
on the POST request).