powershell: add creation_status link to async'ly created resources
When a resource is created asynchronously, we add a creation_status
link to the returned representation so that the client can subsequently
query the status of the resource creation for completion or failure.

For example, when creating a snapshot asynchronously:

  POST /vms/XXX/snapshots HTTP/1.1
  Content-type: application/xml
  <snapshot>
    <description>checkpoint</name>
  </snapshot>

  HTTP/1.1 202 Accepted
  Content-type: application/xml
  <snapshot id="YYY" href="/vms/XXX/snapshots/YYY">
    <description>checkpoint</description>
    <link rel="creation_status"
          href="/vms/XXX/snapshots/YYY/creation_status/ZZZ"/>
    <creation_status>PENDING</creation_status>
    <vm id="XXX" href="/vms/XXX"/>
  </snapshot>

The rel="creation_status" link encapsulates one or more async task IDs
("ZZZ" in the illustration above).