powershell: add CreationResource implementation
This allows the creation_status link for asynchronously created resources
to be queried post-creation, so that the client can determine whether it
completed successfully.

However it relies on a powershell feature (the -CommandTasksIds option to
the Get-TasksStatus cmdlet) that will not be available until RHEVM 2.2.5 is
released.

The Creation representation includes a status (pending, in progress,
complete or failed) and also a fault where appropriate, for example:

  GET /vms/XXX/disks/YYY/creation_status/ZZZ HTTP/1.1
  Accept: application/xml

  HTTP/1.1 200 OK
  Content-type: application/xml
  <creation id="ZZZ" href="/vms/XXX/disks/YYY/creation_status/ZZZ">
    <status>FAILED</status>
    <fault>
      <reason>RHEVM asynchronous task failed</reason>
      <detail>System.Exception: ...</detail>
    </fault>
  </creation>