api: adding schema types for Statistics
Statistics are self-describing, with a unit (bytes, seconds, bytes/s etc.)
and type (guage versus counter), as well as the usual name and decription.

Statistic instances contain a sequence of values, with a type attribute
defined on the sequence to distinguish between integer and decimal data.

Values currently only contain a data-point, but will in the future also
include a timestamp. Note the data-point schema type is always double,
even if the enclosing value is of type integer as opposed to decimal.

So a representation might look something like the following:

  <statistics>
    <statistic id="YYY" href="/vms/XXX/statistics/YYY">
      <name>current.rx</name>
      <description>Receive rate</description>
      <unit>BITS_PER_SECOND</unit>
      <type>GUAGE</type>
      <values type="DECIMAL">
        <value>
          <datum>131072.0<datum>
        </value>
      </values>
    </statistic>
    <statistic id="ZZZ" href="/vms/XXX/statistics/ZZZ">
      <name>errors.total.rx</name>
      <description>Receive errors</description>
      <unit>NONE</unit>
      <type>COUNTER</type>
      <values type="INTEGER">
        <value>
          <datum>1.0<datum>
        </value>
      </values>
    </statistic>
  </statistic>

Acked-by: Mark McLoughlin <markmc@redhat.com>