api: adding Statistics resource interfaces.
Entities exposing statistics implement the MeasurableResource interface.

(I'm not too hot on the DiskResource and NicResource interfaces, as these
slightly undermine the elegance of the generic DeviceResource. However
at least for the DiskResource, this has the advantage of allowing a future
change so that only mutable Disks are AsynchronouslyCreatable, as opposed
to any device type.)

Acked-by: Mark McLoughlin <markmc@redhat.com>