api: adding parent resource links to Statistic schema type
Adding references to Disk, Host, NIC, HostNIC and VM, only one of which
should be set for a given instance.

The parent reference allows the LinkHelper to correctly construct the
Statistic self link.

Also remove a <jaxb:property> customization to avoid an apparent bug
in JAXB codegen that causes all the Statistic child elements to be
munged into a single List<Object>.

Acked-by: Mark McLoughlin <markmc@redhat.com>