api: add <code/> under <event/>
https://fedorahosted.org/rhevm-api/ticket/26

Geert makes a good point that the event code is already a part of the
API through the search syntax. Also, the event codes will not change
from release to release.

No symbolic name will be associated initially with event codes, so it
will be important that the event codes are documented.

Acked-by: Eoghan Glynn <eglynn@redhat.com>