powershell: fix ExpectJ issue causing hangs
https://fedorahosted.org/rhevm-api/ticket/154

ExpectJ isn't properly handling its buffers. It tries to match the
pattern we are looking for on one or multiple full lines of output (to
prevent from incidentally not matching if our pattern spans multiple
reads - note it is assumed that a pattern never contains a
newline). However, contents from a previous read were left in the
buffer, if the subsequent read was less bytes than the current
read. If our end marker was read halfway, and there was a stray
newline in the buffer from a previous read, then ExpectJ would consume
the line but not find the match.

Acked-by: Mark McLoughlin <markmc@redhat.com>
Acked-by: Eoghan Glynn <eglynn@redhat.com>