Simplify the way in which we use the PyXB generated schema.
Previously our approached required us to:

 1. put all mapped types in _schema._mapping_data
 2. bind all ComplexType's to an xs:element in the root of the XMLSchema

The reason for this was that you can't instantiate a ComplexType
directly. A ComplexType can be bound to multiple elements, and so PyXB
doesn't who which tag to use when serializing the class with toxml(). By
requiring all ComplexType's to be bound to top-level elements in the
XMLSchema, we solved the problem as PyXB would generate special "element
constructors" that perform the binding.

This approach however turned out to be quite a maintance overhead,
especially because the number of child-element ComplexType's is growing
(e.g. KSM, TransparentHugepages, etc). It is also only a matter of time
before we get naming conflicts.

The new approach is to only explicitly require us bind ComplexType's to
an element if the element is a top-level element. That list should be
relatively static. For sub-elements, PyXB contains a nifty feature where
an assignment to an element with a complex type will bind the element
automatically. Example:

  vm = schema.new(schema.VM)  # serializes as <vm/>
  os = schema.new(OperatingSystem)  # serializes as <OperatingSystem/>
  vm.os = os  # now it will serialize as <vm><os/></vm>

This patch also merges the lastest schema changes.