Allow specification of the detail level for the response.
The API allows a user to specify the requested detail of a response
entity like this:

  Content-Type application/xml; detail=statistics

This would add the detail "statistics" to the response entity. The
following methods got a 'detail' keyword argument: get(), getall() and
reload(). It passed a string, and that string will be added literally to
the "detail" sub-parameter to Content-Type.