Temporarily use forked version of PyXB
The package of PyXB on the Python Package Index does not work, neither
on Fedora 14 nor on RHEL5. This instructs the installer to use an
alternative package called "pyxbbase," which is simply a copy of
PyXB-base-1.1.2 without any changes but uploaded under a different name.
This allows uses to test out python-rhev and rhevsh simply with
easy_install. When upstream fixes this, we will switch back right away.

Upstream bug report:

http://sourceforge.net/projects/pyxb/forums/forum/956708/topic/4376190