powershell: Optionally parse with a detail set
The rationale behind this, is that sometimes you want a parsed VM set
with certain details, but the powershell backend does not return this.
One example is disks. If you want inline disks and a VM has none, you do
want vm.isSetDisks() to return true so you end up with an empty
collection.

Acked-By: Mark McLoughlin <markmc@redhat.com>