powershell: fix adding hosts (#225)
https://fedorahosted.org/rhevm-api/ticket/225

The current code procudes a powershell scriptlet like this:

  $h = add-host ...; $h| foreach { }

In the case of VMs, the scriptlet looks like:

  $v = add-vm ...; $v; $v| foreach { }

but that is only needed because of async support for VM creation.

We really just want to go back to a PROCESS_VMS which is the same
whether adding or listing VMs.

Reported-by: Geert Jansen <gjansen@redhat.com>
Acked-by: Ori Liel <oliel@redhat.com>