api: re-model host status
A host's <status/> element currently exposes too many fine-grained host
states in the schema. This set of states would be hard to guarantee
future compatibility around, but they are also largely uninteresting
from the perspective of a client that wishes to manipulate a client
according to its state machine.

Reduce the number of states to "down", "active", "inactive", "error" and
"pending approval" e.g.

  GET /rhevm-api/hosts/001 HTTP/1.1
  Accept: application/xml

  HTTP/1.1 200 OK
  Content-Type: application/xml

  <host id="001" href="/rhevm-api/hosts/001">
    ...
    <status>
      <state>error</state>
      <detail>non responsive</detail>
    </status>
    ...
  </host>

More details on the rationale for the change here:

  https://fedorahosted.org/pipermail/rhevm-api/2011-March/001539.html