common: implement mapping host states in search queries
Add some simple code which transforms e.g.

  Hosts: status = active

to:

  Hosts: status = up

i.e. mapping the API state "active" to the backend state "up".

This allows API clients to search using the host states we use in the
API, rather than the backend states.

The transform is based on a very simple regex, but it appears to catch
all the cases we care about. See the test case for examples.