api: use xs:decimal rather than xs:double for statistic values
The decimal type, along with its representation as a string, is much more
suitable to the values we typically see for statistics.

The double type is suitable for performing binary floating point
arithmetic which gives results as approximations, not exact values.

Our units (e.g. percentages or bytes per second) do not require the
large ranges afforded by the floating point numbers, but we do require
the precision.

Care is taken in the code to not convert any value to double before
converting to decimal, so as to avoid the associated loss of precision.

Acked-by: Ori Liel <oliel@redhat.com>