api: add VM <memory_policy/> with minimum guaranteed memory
e.g.

  <vm>
    ...
    <memory_policy>
      <guaranteed>123456789</guaranteed>
    </memory_policy>
    ...
  </vm>

The property is used to help tune a cluster's scheduling policy by
requiring that a host must have the specified amount of unshared
memory available before this VM may be scheduled on that host.

The default value for the property is:

  (vm_mem * 100 / cluster_over_commit_percentage)

i.e. if the cluster is configured to over-commit to 200% then, by
default, VMs require 50% of their memory to be free on a given host
before they can be scheduled onto that host.

When we migrate the VM to a specific host we want to guarantee a min
available memory to avoid a situation where the VM is going to swap
before KSM kicks in.

The issue was that VM after migration consumes almost immediately a lot
of the memory it has allocated for it and does not give KSM a 'chance'
to identify shared pages for aggregation before going to swap (if the
host is on the edge of the available memory).

To avoid such situations the above property was added and now the user
can tune how much memory is required on the host for a VM to migrate to
it.

After the VM is reported UP by the host and we have update statistics on
the host's used memory we don't use this property and the memory is not
really guaranteed.

Acked-by: Mark McLoughlin <markmc@redhat.com>