Implement a "console" command.
This adds  command that will allow you to connect to a console of a VM.
Usage:

 $ console myvm

This supported both VNC and SPICE, currently only on Linux. Of course
you need a VNC viewer (today only tigervnc is supported) or the
SPICE client installed.

Credits to Tomas von Veschler Cox for finding out all the SPICE details
which i took from his spice.py:

 https://code.google.com/p/rhev-api-labs/source/browse/python/spice/spice.py