powershell: lookup datacenter id from storage domain
https://fedorahosted.org/rhevm-api/ticket/217

Based on a patch series from Michael Pasternak.

Assume that an export storage domain can only be attached to a single
datacenter and lookup the datacenter using this scriptlet:

  $datacenter = $null;
  foreach ($dc in select-datacenter) {
    $sds = @();
    try {
      $sds = get-storagedomain -datacenterid $dc.datacenterid
    } catch { }
    foreach ($sd in $sds) {
      if ($sd.storagedomainid -eq $storagedomainid) {
        $datacenter = $dc;
        break
      }
    }
    if ($datacenter -ne $null) { break }
  }

Note, the try/catch is needed because the get-storagedomain command will
throw an exception if it is not attached to the specified data center.