powershell: lookup datacenter from cluster for vm/template import
https://fedorahosted.org/rhevm-api/ticket/217

Based on a patch series from Michael Pasternak.

A datacenter ID must be supplied to the import-vm and import-template
commands, but this is no longer available from the URI of the import
candidate.

For VM/template import, the client must supply a cluster ID or name.
Using that, we can lookup the datacenter by doing:

 select-datacenter -searchtext $("clusters.name = " + $cluster.name)